feat: display-post-in-dashboard
